Move cell tests to extra file
