+++ tests/ids-data-grid/ids-data-grid-cell.spec.ts
@@ -1,0 +1,111 @@
+import { expect } from '@playwright/test';
+import { test } from '../base-fixture';
+
+import IdsDataGrid from '../../src/components/ids-data-grid/ids-data-grid';
+import IdsDataGridCell from '../../src/components/ids-data-grid/ids-data-grid-cell';
+
+test.describe('IdsDataGridCell tests', () => {
+  const url = '/ids-data-grid/columns-formatters.html';
+
+  test.beforeEach(async ({ page }) => {
+    await page.goto(url);
+  });
+
+  test.describe('functionality tests', () => {
+    test('renders rowNumber cells', async ({ page }) => {
+      const cell = await page.locator('ids-data-grid ids-data-grid-cell:nth-child(3)').first();
+      expect(await cell.innerHTML()).toBe(`<span class="text-ellipsis">1</span>`);
+      await page.evaluate(() => {
+        const elem = document.querySelector<IdsDataGridCell>('ids-data-grid ids-data-grid-cell:nth-child(2)');
+        elem?.renderCell();
+      });
+      expect(await cell.innerHTML()).toBe(`<span class="text-ellipsis">1</span>`);
+    });
+
+    test('renders custom formatters cells', async ({ page }) => {
+      const cell = await page.locator('ids-data-grid ids-data-grid-cell:nth-child(24)').first();
+      expect(await cell.innerHTML()).toBe(`<span class="text-ellipsis">Custom: 12.99</span>`);
+      await page.evaluate(() => {
+        const elem = document.querySelector<IdsDataGridCell>('ids-data-grid ids-data-grid-cell:nth-child(24)');
+        elem?.renderCell();
+      });
+      expect(await cell.innerHTML()).toBe(`<span class="text-ellipsis">Custom: 12.99</span>`);
+    });
+
+    test.describe('empty message tests', () => {
+      test.beforeEach(async ({ page }) => {
+        await page.goto('/ids-data-grid/empty-message.html');
+      });
+
+      test('can set empty message description', async ({ page }) => {
+        const str = 'test';
+        const locator = await page.locator('ids-data-grid');
+        expect(await locator.getAttribute('empty-message-description')).toEqual(null);
+        const value = await page.evaluate(() => {
+          const elem = document.querySelector<IdsDataGrid>('ids-data-grid')!;
+          return elem.emptyMessageDescription;
+        });
+        expect(await value).toEqual(null);
+
+        const value2 = await page.evaluate((testStr: string) => {
+          const elem = document.querySelector<IdsDataGrid>('ids-data-grid')!;
+          elem.emptyMessageDescription = testStr;
+          return elem.emptyMessageDescription;
+        }, str);
+        expect(await locator.getAttribute('empty-message-description')).toEqual(str);
+        expect(await value2).toEqual(str);
+
+        const value3 = await page.evaluate(() => {
+          const elem = document.querySelector<IdsDataGrid>('ids-data-grid')!;
+          elem.emptyMessageDescription = '';
+          return elem.emptyMessageDescription;
+        });
+        expect(await locator.getAttribute('empty-message-description')).toEqual(null);
+        expect(await value3).toEqual(null);
+
+        const value4 = await page.evaluate(() => {
+          const elem = document.querySelector('ids-data-grid') as any;
+          elem.emptyMessageDescription = true;
+          return elem.emptyMessageDescription;
+        });
+        expect(await locator.getAttribute('empty-message-description')).toEqual(null);
+        expect(await value4).toEqual(null);
+      });
+
+      test('can set empty message label', async ({ page }) => {
+        const str = 'test';
+        const locator = await page.locator('ids-data-grid');
+        expect(await locator.getAttribute('empty-message-label')).toEqual(null);
+        const value = await page.evaluate(() => {
+          const elem = document.querySelector<IdsDataGrid>('ids-data-grid')!;
+          return elem.emptyMessageLabel;
+        });
+        expect(await value).toEqual(null);
+
+        const value2 = await page.evaluate((testStr: string) => {
+          const elem = document.querySelector<IdsDataGrid>('ids-data-grid')!;
+          elem.emptyMessageLabel = testStr;
+          return elem.emptyMessageLabel;
+        }, str);
+        expect(await locator.getAttribute('empty-message-label')).toEqual(str);
+        expect(await value2).toEqual(str);
+
+        const value3 = await page.evaluate(() => {
+          const elem = document.querySelector<IdsDataGrid>('ids-data-grid')!;
+          elem.emptyMessageLabel = '';
+          return elem.emptyMessageLabel;
+        });
+        expect(await locator.getAttribute('empty-message-label')).toEqual(null);
+        expect(await value3).toEqual(null);
+
+        const value4 = await page.evaluate(() => {
+          const elem = document.querySelector('ids-data-grid') as any;
+          elem.emptyMessageLabel = true;
+          return elem.emptyMessageLabel;
+        });
+        expect(await locator.getAttribute('empty-message-label')).toEqual(null);
+        expect(await value4).toEqual(null);
+      });
+    });
+  });
+});
